+++ itachallenge-challenge/src/main/java/com/itachallenge/App.java
@@ -3,15 +3,9 @@ package com.itachallenge;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-@ComponentScan(basePackages = {
-        "com.itachallenge.challenge",
-        "com.itachallenge.jwtcore"
-})
-
 public class App {
 
     public static void main(String[] args) {